Added: code refactoring for package manger, small code changes in struct

+++ Source/MyProject/Public/PackageDataStruct.h
@@ -44,7 +44,7 @@ struct FConfigPackage
 		int32 Rarity;
 
 	UPROPERTY()
-		int32 PackageWeight;
+		float PackageWeight;
 
 	FConfigPackage() {};
